requirements.txt
astroid==2.15.5
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dill==0.3.6
exceptiongroup==1.1.1
Flask==2.3.2
idna==3.4
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
packaging==23.1
platformdirs==3.5.1
pluggy==1.0.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pyaes==1.6.1
pyasn1==0.5.0
pylint==2.17.4
pyTelegramBotAPI==4.11.0
pytest==7.3.1
python-dotenv==1.0.0
requests==2.30.0
rsa==4.9
telebot==0.0.5
Telethon==1.28.5
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.5.0
urllib3==2.0.2
Werkzeug==2.3.4
wrapt==1.15.0
